A4P2 minor changes to finder-app

+++ finder-app/finder-test.sh
@@ -8,7 +8,9 @@ set -u
 NUMFILES=10
 WRITESTR=AELD_IS_FUN
 WRITEDIR=/tmp/aeld-data
-username=$(cat conf/username.txt)
+username=$(cat /etc/finder-app/conf/username.txt)
+
+OUTPUTFILE=/tmp/assignment4-result.txt
 
 if [ $# -lt 3 ]
 then
@@ -32,7 +34,7 @@ echo "Writing ${NUMFILES} files containing string ${WRITESTR} to ${WRITEDIR}"
 rm -rf "${WRITEDIR}"
 
 # create $WRITEDIR if not assignment1
-assignment=`cat conf/assignment.txt` #*********
+assignment=`cat /etc/finder-app/conf/assignment.txt` 
 
 if [ $assignment != 'assignment1' ]
 then
@@ -50,8 +52,8 @@ then
 fi
 
 echo "Removing the old writer utility and compiling as a native application"
-# make clean
-# make
+make clean
+make
 
 for i in $( seq 1 $NUMFILES)
 do
@@ -60,6 +62,19 @@ do
 done
 
 OUTPUTSTRING=$(./finder.sh "$WRITEDIR" "$WRITESTR")
+
+
+touch $(OUTPUTFILE)
+if [ $? -ne 0 ]
+then
+    echo "$(OUTPUTFILE) could not be created."
+    exit 1
+fi
+
+echo "$(OUTPUTSTRING)" > $(OUTPUTFILE)
+
+
+
 
 # remove temporary directories
 rm -rf /tmp/aeld-data
